Create KernelPlugin from Open API without Kernel instance
